submission/requirements.txt
asgiref==3.11.0
certifi==2025.11.12
charset-normalizer==3.4.4
Django==5.2.8
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
Faker==40.1.0
gunicorn==25.3.0
idna==3.11
mysqlclient==2.2.7
packaging==26.0
pillow==12.1.1
PyJWT==2.12.1
requests==2.32.5
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.6.2
